# rest_service/api/swagger/swagger.yaml
openapi: 3.0.0
info:
  version: 3.0.0
  title: FFAStrans REST API Version 3.0
servers:
  - url: /
  - url: /api/json/v2
paths:
...
  /workflows:
    x-swagger-router-controller: workflows
    get:
      description: ""
      operationId: workflows
      parameters:
        - name: id
          in: query
          description: Workflow ID, if not  specified returns all
          # problem of the old ffastrans api: parameters as part of url (not get params but url) must be required but ffastrans has optional params
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Success
        default:
          description: Error
    post:
      description: "Import Workflow"
      operationId: workflow_post
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                name:
                  type: string
              example: {"workflows": []}
      responses:
        "200":
          description: Success
        default:
          description: Error
  /variables:
    x-swagger-router-controller: variables
    get:
      description: ""
      operationId: variables
      responses:
        "200":
          description: Success
        default:
          description: Error
    delete:
      description: Deletes a variable
      operationId: variables_delete
      parameters:
        - name: name
          in: query
          description: Name of the variable to be deleted
          required: true
          schema:
            type: string
      
      responses:
        "200":
          description: Success
        default:
          description: Error
    post:
      description: "Create/Update Variable"
      operationId: variables_post
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                name:
                  type: string
              example: {"name":"s_example","data":"Example Description/Value"}
      responses:
        "200":
          description: Success
        default:
          description: Error
...